Phase 0: foundation refactors for vanilla-parity roadmap
- Widen chunk block storage to Uint16Array + parallel per-voxel meta array
- Move ITEM ids to 1000+ (frees block ids 78-999); fix hardcoded drop ids
- Expand texture atlas 4x42 -> 8x64 (512 tiles)
- Add dimension descriptors (overworld/nether/end) replacing skyless flag
- Encode block edits as id|meta<<12 with legacy-compatible decode
- Crafting: shape normalization + auto-mirroring, drop hand-written mirrors
- Save migration chain in storage.js (v7 -> v8 item id remap, parked dims)
- Add test/smoke.mjs assertion suite and TESTING.md manual checklist

+++ Minecraft/js/lighting.js
@@ -26,7 +26,9 @@ const localIndex = (x, y, z) => x + CHUNK_SIZE * (z + CHUNK_SIZE * y);
 const idx = (x, y, z) => x + W * (z + W * y);
 
 // Scratch buffers reused across calls (single-threaded main loop / worker).
-let blocks = new Uint8Array(VOL);
+// `blocks` mirrors chunk data (uint16: block ids can exceed 255); light levels
+// stay 0..15 so the sky/blk fields remain bytes.
+let blocks = new Uint16Array(VOL);
 let sky = new Uint8Array(VOL);
 let blk = new Uint8Array(VOL);
 let queue = new Int32Array(VOL);
